pyproject.toml
[tool.poetry]
name = "api-yatube"
version = "0.1.0"
description = ""
authors = ["Марина Ежова <marickyarick@gmail.com>"]
readme = "README.md"
packages = [{include = "api_yatube"}]

[tool.poetry.dependencies]
python = "^3.7"

[tool.poetry.group.dev.dependencies]
Django = "3.2"
djangorestframework = "3.12.4"
asgiref = "3.5.2"
attrs = "22.1.0"
iniconfig = "1.1.1"
packaging = "21.3"
Pillow = "9.3.0"
pluggy = "0.13.1"
py = "1.11.0"
pyparsing = "3.0.9"
pytest = "6.2.4"
pytest-django = "4.4.0"
pytest-pythonpath = "0.7.3"
pytz = "2022.6"
sqlparse = "0.4.3"
toml = "0.10.2"
black = "^22.12.0"
mypy = "^0.991"

[tool.poetry.group.flake.dependencies]
flake8-absolute-import = "^1.0.0.1"
flake8-bugbear = "^22.12.6"
flake8-broken-line = "^0.6.0"
flake8-cognitive-complexity = "^0.1.0"
flake8-commas = "^2.1.0"
flake8-comprehensions = "^3.10.1"
flake8-eradicate = "^1.4.0"
flake8-fixme = "^1.1.1"
flake8-isort = "^5.0.3"
flake8-multiline-containers = "^0.0.19"
flake8-mutable = "^1.2.0"
flake8-pep3101 = "^2.0.0"
flake8-pie = "^0.16.0"
flake8-print = "^5.0.0"
flake8-printf-formatting = "^1.1.2"
flake8-quotes = "^3.3.1"
flake8-return = "^1.2.0"
flake8-simplify = "^0.19.3"
flake8-todo = "^0.7"
pep8-naming = "^0.13.2"
flake8-variables-names = "^0.0.5"
flake8-use-fstring = "^1.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
